backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.4
django-allauth==0.58.2
django-auth-style==0.8.1
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
idna==3.6
oauthlib==3.2.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
stripe==7.7.0
typing_extensions==4.8.0
urllib3==2.1.0
